Started to move providers to new namespace
