+++ src/Common/Collection.php
