+++ src/Common/Dumper/GeoArray.php
