+++ src/Common/Dumper/GeoJson.php
