+++ src/Common/Exception/ChainZeroResults.php
