+++ src/Common/Exception/CollectionIsEmpty.php
