+++ src/Common/Exception/Exception.php
