+++ src/Common/Exception/HttpError.php
