+++ src/Common/Exception/InvalidServerResponse.php
