+++ src/Common/Exception/ProviderNotRegistered.php
