+++ src/Common/Exception/UnsupportedOperation.php
