+++ src/Common/Formatter/StringFormatter.php
