+++ src/Common/Provider/AbstractHttpProvider.php
