+++ src/Common/ProviderAggregator.php
