+++ src/Provider/Chain/composer.json
@@ -1,0 +1,36 @@
+{
+    "name": "geocoder-php/chain-provider",
+    "type": "library",
+    "description": "Geocoder chain adapter",
+    "keywords": [],
+    "homepage": "http://geocoder-php.org/Geocoder/",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "William Durand",
+            "email": "william.durand1@gmail.com"
+        }
+    ],
+    "require": {
+        "php": "^5.5 || ^7.0",
+        "willdurand/geocoder": "^4.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^4.8",
+        "php-http/message": "^1.0",
+        "php-http/guzzle6-adapter": "^1.0",
+        "symfony/phpunit-bridge": "^3.2"
+    },
+    "autoload": {
+        "psr-4": { "Geocoder\\Provider\\Chain\\": "" },
+        "exclude-from-classmap": [
+            "/Tests/"
+        ]
+    },
+    "minimum-stability": "dev",
+    "extra": {
+        "branch-alias": {
+            "dev-master": "4.0-dev"
+        }
+    }
+}
